Move row save to sdk
